requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Login>=0.6
Flask-Babel>=4.0
Flask-WTF>=1.2
python-dotenv>=1.0
gunicorn>=21.0
Werkzeug>=3.0
